fix for script stall when starting height is low
logic to the add piece function which should hopefully prevent pieces from going below ground and prevents the script from stalling. If the current height of the piece is found to be below zero, before it adds the piece it overrides the currentHeight value. starting height values below 1 will still stall the script